# swagger-codegen/jaxrs-cxf-cdi/swagger.json
{
  "swagger" : "2.0",
  "info" : {
...
    "version" : "1.0.0",
    "title" : "Samsara API"
  },
  "host" : "api.samsara.com",
  "basePath" : "/v1",
...
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/tags" : {
      "get" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags",
        "description" : "Fetch all of the tags for a group.",
        "operationId" : "GetAllTags",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "group_id",
          "in" : "query",
          "description" : "Optional group ID if the organization has multiple groups (uncommon).",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "List of tags.",
            "schema" : {
              "type" : "object",
              "properties" : {
                "tags" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Tag"
                  }
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags",
        "description" : "Create a new tag for the group.",
        "operationId" : "CreateTag",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "tagCreateParams",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TagCreate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Newly created tag object, including the new tag ID.",
            "schema" : {
              "$ref" : "#/definitions/Tag"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tags/{tag_id}" : {
      "get" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags/{tag_id:[0-9]+}",
        "description" : "Fetch a tag by id.",
        "operationId" : "getTagById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tag_id",
          "in" : "path",
          "description" : "ID of the tag.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "The tag corresponding to tag_id.",
            "schema" : {
              "$ref" : "#/definitions/Tag"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags/{tag_id:[0-9]+}",
        "description" : "Update a tag with a new name and new members. This API call would replace all old members of a tag with new members specified in the request body. To modify only a few devices associated with a tag use the PATCH endpoint.",
        "operationId" : "updateTagById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tag_id",
          "in" : "path",
          "description" : "ID of the tag.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "updateTagParams",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TagUpdate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The updated tag data.",
            "schema" : {
              "$ref" : "#/definitions/Tag"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags/{tag_id:[0-9]+}",
        "description" : "Permanently deletes a tag.",
        "operationId" : "deleteTagById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tag_id",
          "in" : "path",
          "description" : "ID of the tag.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted the tag. No response body is returned."
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Tags" ],
        "summary" : "/tags/{tag_id:[0-9]+}",
        "description" : "Add or delete specific members from a tag, or modify the name of a tag.",
        "operationId" : "modifyTagById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tag_id",
          "in" : "path",
          "description" : "ID of the tag.",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "tagModifyParams",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TagModify"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The updated tag data.",
            "schema" : {
              "$ref" : "#/definitions/Tag"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ErrorResponse" : {
      "type" : "string",
      "description" : "Error message describing why the request failed."
    },
...
    "Tag" : {
      "type" : "object",
      "required" : [ "id", "name" ],
      "properties" : {
        "addresses" : {
          "type" : "array",
          "description" : "The addresses that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedAddress"
          }
        },
        "assets" : {
          "type" : "array",
          "description" : "The assets that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedAsset"
          }
        },
        "drivers" : {
          "type" : "array",
          "description" : "The drivers that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedDriver"
          }
        },
        "groupId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 2348,
          "description" : "The GroupID that this tag belongs to."
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 12345,
          "description" : "The ID of this tag."
        },
        "machines" : {
          "type" : "array",
          "description" : "The machines that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedMachine"
          }
        },
        "name" : {
          "type" : "string",
          "example" : "Broken Vehicles",
          "description" : "Name of this tag."
        },
        "parentTagId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 8389,
          "description" : "If this tag is part a hierarchical tag tree as a child tag, the parentTagId is the ID of this tag's parent tag."
        },
        "sensors" : {
          "type" : "array",
          "description" : "The sensors that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedSensor"
          }
        },
        "vehicles" : {
          "type" : "array",
          "description" : "The vehicles that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedVehicle"
          }
        }
      }
    },
    "TagCreate" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "assets" : {
          "type" : "array",
          "description" : "The assets that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedAssetBase"
          }
        },
        "drivers" : {
          "type" : "array",
          "description" : "The drivers that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedDriverBase"
          }
        },
        "machines" : {
          "type" : "array",
          "description" : "The machines that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedMachineBase"
          }
        },
        "name" : {
          "type" : "string",
          "example" : "Broken Vehicles",
          "description" : "Name of this tag."
        },
        "parentTagId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 8389,
          "description" : "If this tag is part a hierarchical tag tree as a child tag, the parentTagId is the ID of this tag's parent tag."
        },
        "sensors" : {
          "type" : "array",
          "description" : "The sensors that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedSensorBase"
          }
        },
        "vehicles" : {
          "type" : "array",
          "description" : "The vehicles that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedVehicleBase"
          }
        }
      }
    },
...
    "TagModify" : {
      "type" : "object",
      "properties" : {
        "add" : {
          "type" : "object",
          "description" : "Specify devices, etc. that should be added to the tag.",
          "properties" : {
            "assets" : {
              "type" : "array",
              "description" : "The assets to be added to this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedAssetBase"
              }
            },
            "drivers" : {
              "type" : "array",
              "description" : "The drivers to be added to this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedDriverBase"
              }
            },
            "machines" : {
              "type" : "array",
              "description" : "The machines to be added to this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedMachineBase"
              }
            },
            "sensors" : {
              "type" : "array",
              "description" : "The sensors to be added to this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedSensorBase"
              }
            },
            "vehicles" : {
              "type" : "array",
              "description" : "The vehicles to be added to this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedVehicleBase"
              }
            }
          }
        },
        "delete" : {
          "type" : "object",
          "description" : "Specify devices, etc. that should be removed from the tag.",
          "properties" : {
            "assets" : {
              "type" : "array",
              "description" : "The assets to be removed from this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedAssetBase"
              }
            },
            "drivers" : {
              "type" : "array",
              "description" : "The drivers to be removed from this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedDriverBase"
              }
            },
            "machines" : {
              "type" : "array",
              "description" : "The machines to be removed from this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedMachineBase"
              }
            },
            "sensors" : {
              "type" : "array",
              "description" : "The sensors to be removed from this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedSensorBase"
              }
            },
            "vehicles" : {
              "type" : "array",
              "description" : "The vehicles to be removed from this tag.",
              "items" : {
                "$ref" : "#/definitions/TaggedVehicleBase"
              }
            }
          }
        },
        "name" : {
          "type" : "string",
          "example" : "Broken Vehicles",
          "description" : "Updated name of this tag."
        },
        "parentTagId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 8389,
          "description" : "If this tag is part a hierarchical tag tree as a child tag, the parentTagId is the ID of this tag's parent tag."
        }
      }
    },
    "TagUpdate" : {
      "type" : "object",
      "properties" : {
        "assets" : {
          "type" : "array",
          "description" : "The assets that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedAssetBase"
          }
        },
        "drivers" : {
          "type" : "array",
          "description" : "The drivers that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedDriverBase"
          }
        },
        "machines" : {
          "type" : "array",
          "description" : "The machines that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedMachineBase"
          }
        },
        "name" : {
          "type" : "string",
          "example" : "Broken Vehicles",
          "description" : "Updated name of this tag."
        },
        "parentTagId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 8389,
          "description" : "If this tag is part a hierarchical tag tree as a child tag, the parentTagId is the ID of this tag's parent tag."
        },
        "sensors" : {
          "type" : "array",
          "description" : "The sensors that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedSensorBase"
          }
        },
        "vehicles" : {
          "type" : "array",
          "description" : "The vehicles that belong to this tag.",
          "items" : {
            "$ref" : "#/definitions/TaggedVehicleBase"
          }
        }
      }
    },
    "TaggedAddress" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedAddressBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "Truck Lot 345",
            "description" : "Name of the address being tagged."
          }
        }
      } ]
    },
    "TaggedAddressBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 345,
          "description" : "The ID of the address being tagged."
        }
      }
    },
    "TaggedAsset" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedAssetBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "Trailer 789",
            "description" : "Name of the Asset being tagged."
          }
        }
      } ]
    },
    "TaggedAssetBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 789,
          "description" : "The ID of the Asset being tagged."
        }
      }
    },
    "TaggedDriver" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedDriverBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "John Smith",
            "description" : "Name of the Driver being tagged."
          }
        }
      } ]
    },
    "TaggedDriverBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 456,
          "description" : "The ID of the Driver being tagged."
        }
      }
    },
    "TaggedMachine" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedMachineBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "Vibration Monitor 567",
            "description" : "Name of the Machine being tagged."
          }
        }
      } ]
    },
    "TaggedMachineBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 567,
          "description" : "The ID of the Machine being tagged."
        }
      }
    },
    "TaggedSensor" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedSensorBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "Temperature Sensor 345",
            "description" : "Name of the Sensor being tagged."
          }
        }
      } ]
    },
    "TaggedSensorBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 345,
          "description" : "The ID of the Sensor being tagged."
        }
      }
    },
    "TaggedVehicle" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TaggedVehicleBase"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "Heavy Duty 123",
            "description" : "Name of the Vehicle being tagged."
          }
        }
      } ]
    },
    "TaggedVehicleBase" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 123,
          "description" : "The ID of the Vehicle being tagged."
        }
      }
    },
...
  },
...
}